pyproject.toml
[project]
name = "kafka-chat"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "vhzkclq0705", email = "vhzkclq0705@naver.com"},
]
dependencies = ["tqdm>=4.67.1", "kafka-python>=2.1.5", "confluent-kafka>=2.10.0", "prompt-toolkit>=3.0.51", "textual>=3.1.1", "textual-dev>=1.7.0"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.scripts]
chatpro = "kafka_producer:main"

[tool.pdm]
distribution = true
